Move cache options to a separate file
